requirements.txt
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.30.0
requests>=2.31.0
geopy>=2.4.1
google-generativeai>=0.3.2
openai>=1.10.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
sshtunnel>=0.4.0
geopandas>=0.14.3
shapely>=2.0.3
pyarrow>=15.0.0
openpyxl>=3.1.2
